pyproject.toml
[project]
name = "fastapi-htmx-login"
version = "0.1.0"
description = "A FastAPI authentication demo featuring password login, magic link login, and password reset using HTMX for seamless frontend interactions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.35.0",
    "python-jose[cryptography]>=3.3.0",
    "jinja2>=3.1.0",
    "pydantic[email]>=2.0.0",
    "emails>=0.6",
]

[project.urls]
Homepage = "https://github.com/rebeling/fastapi-htmx-login"
Repository = "https://github.com/rebeling/fastapi-htmx-login"
Issues = "https://github.com/rebeling/fastapi-htmx-login/issues"

[tool.ruff]
# Keep fast, strict enough defaults for a small demo repo
target-version = "py311"
line-length = 100
exclude = [
  ".venv",
  "app/cognito/email-templates/node_modules",
  "app/cognito/email-templates/build",
  "__pycache__",
]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.uv]
dev-dependencies = [
  "ruff>=0.5.5",
  "pytest>=7.4",
  "httpx>=0.24",
  "pytest-cov>=4.1",
]
